// client/src/context/action.ts
/* eslint-disable import/prefer-default-export */
enum ActionTypes {
  HANDLE_CHANGE = 'HANDLE_CHANGE',
  DISPLAY_ALERT = 'DISPLAY_ALERT',
  CLEAR_ALERT = 'CLEAR_ALERT',
  CLEAR_PIRATE_VALUE = 'CLEAR_PIRATE_VALUE',

  SETUP_USER_BEGIN = 'SETUP_USER_BEGIN',
  SETUP_USER_SUCCESS = 'SETUP_USER_SUCCESS',
  SETUP_USER_ERROR = 'SETUP_USER_ERROR',
  LOGOUT_USER = 'LOGOUT_USER',

  CREATE_PIRATE_BEGIN = 'CREATE_PIRATE_BEGIN',
  CREATE_PIRATE_SUCCESS = 'CREATE_PIRATE_SUCCESS',
  CREATE_PIRATE_ERROR = 'CREATE_PIRATE_ERROR',

  GET_ALL_PIRATES_BEGIN = 'GET_ALL_PIRATES_BEGIN',
  GET_ALL_PIRATES_SUCCESS = 'GET_ALL_PIRATES_SUCCESS',
  GET_ALL_PIRATES_ERROR = 'GET_ALL_PIRATES_ERROR',

  GET_SINGLE_PIRATE_BEGIN = 'GET_SINGLE_PIRATE_BEGIN',
  GET_SINGLE_PIRATE_SUCCESS = 'GET_SINGLE_PIRATE_SUCCESS',
  GET_SINGLE_PIRATE_ERROR = 'GET_SINGLE_PIRATE_ERROR',

  DELETE_PIRATE_BEGIN = 'DELETE_PIRATE_BEGIN',
  DELETE_PIRATE_SUCCESS = 'DELETE_PIRATE_SUCCESS',
  DELETE_PIRATE_ERROR = 'DELETE_PIRATE_ERROR',

  UPDATE_PIRATE_BEGIN = 'UPDATE_PIRATE_BEGIN',
  UPDATE_PIRATE_SUCCESS = 'UPDATE_PIRATE_SUCCESS',
  UPDATE_PIRATE_ERROR = 'UPDATE_PIRATE_ERROR',
}

export default ActionTypes;